style: Creo el componente NavBar para que contenga los links y actue como header [Leandro Bejarano]
